feat&fix #time 30m #comment implement like & comment controller, fixing query sql, route, blog|comment|like interfaces, blog services

+++ service/blog.go
@@ -20,15 +20,15 @@ type (
 		GetAllBlogSvc(page int, size int, order string, field string, search string, filter model.FilterBlogRequest) ([]model.BlogResponse,*model.Metadata,error)
 		GetBlogBySlugSvc(slug string) (*model.BlogResponse,error)
 		UpdateBlogSvc(id int, req model.UpdateBlogRequest, updatedBy string) error
-		DeleteBlogSvc(id int) error
+		DeleteBlogSvc(id int,userID int) error
 
 		CreateCommentSvc(blogID int,req model.CommentRequest,createdBy string) error
 		UpdateCommentSvc(id int, req model.CommentRequest, updatedBy string) error
 		GetCommentsByBlogSvc(blogID int,page int, size int, order string, field string) ([]model.ViewCommentResponse,*model.Metadata,error)
-		DeleteCommentSvc(blogID int, commentID int) error
+		DeleteCommentSvc(blogID int, commentID int, userID int) error
 
 		CreateLikeSvc(blogID int,req model.LikeRequest,createdBy string) error
-		DeleteLikeSvc(blogID int,likeID int) error
+		DeleteLikeSvc(blogID int,likeID int, userID int) error
 	}
 
 	// BlogRepository is an app blog struct that consists of all the dependencies needed for blog service
@@ -97,7 +97,7 @@ func (bs *BlogService) GetAllBlogSvc(page int, size int, order string, field str
 		d.UpdatedAt = r.UpdatedAt
 		d.UpdatedBy = r.UpdatedBy
 
-		if len(r.Comments) > 1 {
+		if len(r.Comments) > 0 {
 			for c := range r.Comments{
 				rc := r.Comments[c]
 				if rc.Valid {
@@ -107,7 +107,7 @@ func (bs *BlogService) GetAllBlogSvc(page int, size int, order string, field str
 		}
 		
 
-		if len(r.Tags) > 1 {
+		if len(r.Tags) > 0 {
 			for t := range r.Tags{
 				rt := r.Tags[t]
 				if rt.Valid {
@@ -116,7 +116,7 @@ func (bs *BlogService) GetAllBlogSvc(page int, size int, order string, field str
 			}
 		}
 
-		if len(r.Likes) > 1 {
+		if len(r.Likes) > 0 {
 			for l := range r.Likes{
 				rl := r.Likes[l]
 				if rl.Valid {
@@ -156,7 +156,7 @@ func (bs *BlogService) GetBlogBySlugSvc(slug string) (*model.BlogResponse,error)
 	data.UpdatedAt = r.UpdatedAt
 	data.UpdatedBy = r.UpdatedBy
 
-	if len(r.Comments) > 1 {
+	if len(r.Comments) > 0 {
 		for c := range r.Comments{
 			rc := r.Comments[c]
 			if rc.Valid {
@@ -166,7 +166,7 @@ func (bs *BlogService) GetBlogBySlugSvc(slug string) (*model.BlogResponse,error)
 	}
 	
 
-	if len(r.Tags) > 1 {
+	if len(r.Tags) > 0 {
 		for t := range r.Tags{
 			rt := r.Tags[t]
 			if rt.Valid {
@@ -175,7 +175,7 @@ func (bs *BlogService) GetBlogBySlugSvc(slug string) (*model.BlogResponse,error)
 		}
 	}
 
-	if len(r.Likes) > 1 {
+	if len(r.Likes) > 0 {
 		for l := range r.Likes{
 			rl := r.Likes[l]
 			if rl.Valid {
@@ -189,14 +189,14 @@ func (bs *BlogService) GetBlogBySlugSvc(slug string) (*model.BlogResponse,error)
 
 // UpdateBlogSvc service layer for handling updating a blog by ID
 func (bs *BlogService) UpdateBlogSvc(id int, req model.UpdateBlogRequest, updatedBy string) error {
-	_,err := bs.BlogRepo.GetByID(id)
+	currentBlog,err := bs.BlogRepo.GetByID(id)
 	if err != nil {
 		if err == pgx.ErrNoRows {
 			return model.ErrBlogNotFound
 		}
 	}
 
-	blogMap,err := validateUpdateBlogRequest(&req)
+	blogMap,err := validateUpdateBlogRequest(currentBlog,&req)
 	if err != nil {
 		return err
 	}
@@ -210,12 +210,16 @@ func (bs *BlogService) UpdateBlogSvc(id int, req model.UpdateBlogRequest, update
 }
 
 // DeleteBlogSvc service layer for handling deleting a blog by ID
-func (bs *BlogService) DeleteBlogSvc(id int) error {
-	_,err := bs.BlogRepo.GetByID(id)
+func (bs *BlogService) DeleteBlogSvc(id int,userID int) error {
+	currentBlog,err := bs.BlogRepo.GetByID(id)
 	if err != nil {
 		if err == pgx.ErrNoRows {
 			return model.ErrBlogNotFound
 		}
+	}
+
+	if currentBlog.UserID != userID {
+		return model.ErrForbiddenDelete
 	}
 
 	err = bs.BlogRepo.DeleteByID(id)
@@ -263,7 +267,7 @@ func (bs *BlogService) CreateCommentSvc(blogID int, req model.CommentRequest,cre
 
 // UpdateCommentSvc service layer for handling updating comment by id
 func (bs *BlogService) UpdateCommentSvc(id int, req model.CommentRequest,updatedBy string) error {
-	_,err := bs.CommentRepo.GetByID(id)
+	currentComment,err := bs.CommentRepo.GetByID(id)
 	if err != nil {
 		if err == pgx.ErrNoRows {
 			return model.ErrCommentNotFound
@@ -272,7 +276,7 @@ func (bs *BlogService) UpdateCommentSvc(id int, req model.CommentRequest,updated
 		return err
 	}
 
-	commentMap,err := validateUpdateCommentRequest(&req)
+	commentMap,err := validateUpdateCommentRequest(currentComment,&req)
 	if err != nil {
 		return err
 	}
@@ -319,7 +323,7 @@ func (bs *BlogService) GetCommentsByBlogSvc(blogID int, page int, size int, orde
 }
 
 // DeleteCommentSvc service layer for handling deleting comment with id
-func (bs *BlogService) DeleteCommentSvc(blogID int, commentID int) error {
+func (bs *BlogService) DeleteCommentSvc(blogID int, commentID int, userID int) error {
 	_,err := bs.BlogRepo.GetByID(blogID)
 	if err != nil {
 		if err == pgx.ErrNoRows {
@@ -329,13 +333,17 @@ func (bs *BlogService) DeleteCommentSvc(blogID int, commentID int) error {
 		return err
 	}
 
-	_,err = bs.CommentRepo.GetByID(commentID)
+	currentComment,err := bs.CommentRepo.GetByID(commentID)
 	if err != nil {
 		if err == pgx.ErrNoRows {
 			return model.ErrCommentNotFound
 		}
 
 		return err
+	}
+
+	if currentComment.UserID != userID {
+		return model.ErrForbiddenDelete
 	}
 
 	bs.Mutex.Lock()
@@ -368,18 +376,28 @@ func (bs *BlogService) CreateLikeSvc(blogID int, req model.LikeRequest,createdBy
 		return err
 	}
 
-	bs.Mutex.Lock()
-	err = bs.LikeRepo.Create(blogID,req,createdBy)
+	_,err = bs.LikeRepo.GetByUserID(req.UserID)
 	if err != nil {
+		if err == pgx.ErrNoRows {
+		
+			bs.Mutex.Lock()
+			err = bs.LikeRepo.Create(blogID,req,createdBy)
+			if err != nil {
+				return err
+			}
+			defer bs.Mutex.Unlock()
+
+			return nil
+		}
+
 		return err
 	}
-	defer bs.Mutex.Unlock()
 
-	return nil
+	return model.ErrAlreadyLikeBlog
 }
 
 // DeleteLikeSvc service layer for handling deleting like with id
-func (bs *BlogService) DeleteLikeSvc(blogID int,likeID int) error {
+func (bs *BlogService) DeleteLikeSvc(blogID int,likeID int, userID int) error {
 	_,err := bs.BlogRepo.GetByID(blogID)
 	if err != nil {
 		if err == pgx.ErrNoRows {
@@ -389,13 +407,17 @@ func (bs *BlogService) DeleteLikeSvc(blogID int,likeID int) error {
 		return err
 	}
 
-	_,err = bs.LikeRepo.GetByID(likeID)
+	currentLike,err := bs.LikeRepo.GetByID(likeID)
 	if err != nil {
 		if err == pgx.ErrNoRows {
 			return model.ErrLikeNotFound
 		}
 
 		return err
+	}
+
+	if currentLike.UserID != userID {
+		return model.ErrForbiddenDelete
 	}
 
 	bs.Mutex.Lock()
@@ -430,7 +452,7 @@ func validateCreateBlogRequest(req *model.CreateBlogRequest) error {
 }
 
 // validateCreateBlogRequest responsible to validating update blog request
-func validateUpdateBlogRequest(req *model.UpdateBlogRequest) (map[string]interface{},error) {
+func validateUpdateBlogRequest(blog *model.ViewBlogResponse,req *model.UpdateBlogRequest) (map[string]interface{},error) {
 	blogMap := make(map[string]interface{})
 
 	if req.Title == "" || req.Body == "" || req.Footer == "" {
@@ -447,6 +469,10 @@ func validateUpdateBlogRequest(req *model.UpdateBlogRequest) (map[string]interfa
 
 	if len(req.Footer) < 5 {
 		return nil,model.ErrInvalidRequest
+	}
+
+	if req.UserID != blog.UserID {
+		return nil,model.ErrForbiddenUpdate
 	}
 
 	req.Slug = slug.Make(req.Title)
@@ -470,11 +496,15 @@ func validateCreateCommentRequest(req *model.CommentRequest) error {
 }
 
 // validateUpdateCommentRequest responsible to validating update comment request
-func validateUpdateCommentRequest(req *model.CommentRequest)(map[string]interface{},error) { 
+func validateUpdateCommentRequest(comment *model.ViewCommentResponse,req *model.CommentRequest)(map[string]interface{},error) { 
 	commentMap := make(map[string]interface{})
 
 	if len(req.Comment) < 5 {
 		return nil,model.ErrInvalidRequest
+	}
+
+	if comment.UserID != req.UserID {
+		return nil,model.ErrForbiddenUpdate
 	}
 
 	commentMap["comment"] = req.Comment
